fix(hosting-overview-refinements): move Overview component to more appropriate place
