+++ client/hosting/overview/components/hosting-overview.tsx
@@ -2,10 +2,10 @@ import config from '@automattic/calypso-config';
 import { translate } from 'i18n-calypso';
 import { FC } from 'react';
 import NavigationHeader from 'calypso/components/navigation-header';
-import ActiveDomainsCard from 'calypso/hosting-overview/components/active-domains-card';
-import PlanCard from 'calypso/hosting-overview/components/plan-card';
-import QuickActionsCard from 'calypso/hosting-overview/components/quick-actions-card';
-import SiteBackupCard from 'calypso/my-sites/hosting/site-backup-card';
+import ActiveDomainsCard from 'calypso/hosting/overview/components/active-domains-card';
+import PlanCard from 'calypso/hosting/overview/components/plan-card';
+import QuickActionsCard from 'calypso/hosting/overview/components/quick-actions-card';
+import SiteBackupCard from 'calypso/hosting/overview/components/site-backup-card';
 import SupportCard from 'calypso/my-sites/hosting/support-card';
 import { isNotAtomicJetpack } from 'calypso/sites-dashboard/utils';
 import { useSelector } from 'calypso/state';
