+++ client/hosting/overview/components/plan-card.tsx
@@ -17,9 +17,9 @@ import PlanStorage from 'calypso/blocks/plan-storage';
 import QuerySitePlans from 'calypso/components/data/query-site-plans';
 import { HostingCard, HostingCardLinkButton } from 'calypso/components/hosting-card';
 import { useLocalizedMoment } from 'calypso/components/localized-moment';
-import { PlanBandwidth } from 'calypso/hosting-overview/components/plan-bandwidth';
-import { PlanSiteVisits } from 'calypso/hosting-overview/components/plan-site-visits';
-import PlanStorageBar from 'calypso/hosting-overview/components/plan-storage-bar';
+import { PlanBandwidth } from 'calypso/hosting/overview/components/plan-bandwidth';
+import { PlanSiteVisits } from 'calypso/hosting/overview/components/plan-site-visits';
+import PlanStorageBar from 'calypso/hosting/overview/components/plan-storage-bar';
 import { isPartnerPurchase, purchaseType } from 'calypso/lib/purchases';
 import useCheckPlanAvailabilityForPurchase from 'calypso/my-sites/plans-features-main/hooks/use-check-plan-availability-for-purchase';
 import { getManagePurchaseUrlFor } from 'calypso/my-sites/purchases/paths';
