+++ client/hosting/overview/controller.tsx
@@ -1,5 +1,5 @@
 import { Context as PageJSContext } from '@automattic/calypso-router';
-import HostingOverview from 'calypso/hosting-overview/components/hosting-overview';
+import HostingOverview from 'calypso/hosting/overview/components/hosting-overview';
 import HostingActivate from 'calypso/my-sites/hosting/hosting-activate';
 import Hosting from 'calypso/my-sites/hosting/main';
